requirements.txt
click>=8.0.1
alabaster>=0.7.12
setuptools>=57.2.0
PyHamcrest~=2.0.2
